+++ docs/changelog.md
@@ -14,6 +14,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Handle bad database encoding.
 - Rename `ldap` to `ldapsearch` in `sync_map`.
 - Ignore spurious ldapsearch YAML parameter.
+- Implement STARTTLS behaviour.
 
 
 # ldap2pg 5.6